inserting new comments in the code

+++ BullCowGame/FBullCowGame.cpp
@@ -4,12 +4,12 @@
 #include <map>
 #include <time.h>
 
-// adaptando a sintaxe utilizad na Unreal Engine
+// to make syntax Unreal friendly
 #define TMap std::map
 using FString = FString;
 using int32 = int;
 
-// contrutor padrão
+// default constructor
 FBullCowGame::FBullCowGame() { 
 	Reset();
 }
@@ -26,7 +26,7 @@ int32 FBullCowGame::GetMaxTries() const {
 }
 
 void FBullCowGame::Reset() {
-	MyHiddenWord = "planet";
+	MyHiddenWord = "planet"; // this MUST be an isogram
 	MyCurrentTry = 1;
 	MyGameLevel = 2;
 	bGameIsWon = false;
@@ -34,27 +34,28 @@ void FBullCowGame::Reset() {
 }
 
 EGuessStatus FBullCowGame::CheckGuesValidity(FString Guess) const {
-	if (!IsIsogram(Guess)) { // verifica se o palpite não é um ISOGRAMA	
+	if (!IsIsogram(Guess)) { // if the guess isn't an ISOGRAM	
 		return EGuessStatus::Not_Isogram;
-	} else if (!IsLowerCase(Guess)) { // verifica se todas as letras são minusculas
+	} else if (!IsLowerCase(Guess)) { // if the guess isn't all lowercase
 		return EGuessStatus::Not_Lowercase; 
-	} else if (Guess.length() != GetHiddenWordLength()) { // verifica o tamanho do palpite 
+	} else if (Guess.length() != GetHiddenWordLength()) { // if the guess length is wrong 
 		return EGuessStatus::Wrong_Length;
 	} else {
 		return EGuessStatus::OK;
 	}
 }
 
-// escolhe aleatóricamente uma palavra de acordo com o level seleciondo
+// choose randomly a word from one of arrays according to level selected by player
 void FBullCowGame::SetHiddenWordByLevelDifficulty()
 {
-	FString EasyDifficultyWords[10] = { "ant", "sky", "pie", "bat", "owl", "bird", "goal", "dice", "fire", "tomb" };
-	FString NormalDifficultyWords[10] = { "planet", "black", "final", "light", "horse", "bridge", "monkey", "gremio", "dragon", "square" };
-	FString HardDifficultyWords[5] = { "pikachu", "soldier", "special", "kremlin", "banquet" };
+	FString EasyDifficultyWords[10] = { "ant", "sky", "pie", "bat", "owl", "bird", "goal", "dice", "fire", "tomb" }; // easy words array
+	FString NormalDifficultyWords[10] = { "planet", "black", "final", "light", "horse", "bridge", "monkey", "gremio", "dragon", "square" }; // normal words array
+	FString HardDifficultyWords[5] = { "pikachu", "soldier", "special", "kremlin", "banquet" }; // hard words array
 
 	int32 MySelectedLevel = GetMyGameLevel();
-	srand(time(NULL));
+	srand((unsigned) time(NULL)); // set the seed based in time to generate always a different seed to prevent generate the same random number
 
+	// check the level and get a word from array position of generated random number
 	if (MySelectedLevel == 1) {
 		this->MyHiddenWord = EasyDifficultyWords[rand() % 10];
 	} else if (MySelectedLevel == 2) {
@@ -64,20 +65,25 @@ void FBullCowGame::SetHiddenWordByLevelDifficulty()
 	}
 }
 
+// receives a VALID guess, incriments turn, and returns count
 FBullCowCount FBullCowGame::SubmitValidGuess(FString Guess) {
 	MyCurrentTry++;
 	FBullCowCount BullCowCount;
-	int32 WordLength = MyHiddenWord.length();
+	int32 WordLength = MyHiddenWord.length(); // assuming same length as guess
 
-	// intera sobre as letras da palavra secreta
+	// loop through all letters in the hidden word
 	for (int32 MHWChar = 0; MHWChar < WordLength; MHWChar++) {
-		//compara as letras da palavra secreta com o palpite do usuário
+		// compare letters against the guess
 		for (int32 GChar = 0; GChar < WordLength; GChar++) {
-			if (Guess[GChar] == MyHiddenWord[MHWChar]) { // se a letra confere
-				if (MHWChar == GChar) { // se a letra está na posição correta
-					BullCowCount.Bulls++; // incrementa Bulls
+			// if they match then
+			if (Guess[GChar] == MyHiddenWord[MHWChar]) { 
+				// if they're in the same place
+				if (MHWChar == GChar) { 
+					// incriment bulls
+					BullCowCount.Bulls++; 
 				} else {
-					BullCowCount.Cows++; // incrementa Cows
+					// must be a cow
+					BullCowCount.Cows++; 
 				}
 			}
 		}
@@ -91,32 +97,36 @@ FBullCowCount FBullCowGame::SubmitValidGuess(FString Guess) {
 	return BullCowCount;
 }
 
+// Set the game level based in player choice
 void FBullCowGame::SetMyGameLevel(int32 MyGameLevel) {
 	this->MyGameLevel = MyGameLevel;
 }
 
 bool FBullCowGame::IsIsogram(FString Word) const {
+	// treat 0 and 1 letter words as isograms
 	if (Word.length() <= 1) {
 		return true;
 	}
 
 	TMap<char, bool> LetterSeen;
+	// loop through all letters of the word
 	for (auto Letter : Word) {
-		Letter = tolower(Letter);
+		Letter = tolower(Letter); // handle mixed case
+		// if the letter is in the map
 		if (LetterSeen[Letter]) {
-			return false;
+			return false; // we do NOT have an isogram
 		} else {
-			LetterSeen[Letter] = true;
+			LetterSeen[Letter] = true; // add the letter to the map
 		}
 
 	}
-  	return true;
+  	return true; // for example in cases where /0 is entered
 }
 
 bool FBullCowGame::IsLowerCase(FString Word) const
 {
 	for (auto Letter : Word) {
-		if (!islower(Letter)) {
+		if (!islower(Letter)) { // if not a lowercase letter
 			return false;
 		}
 	}
